Seperated subspace functions from counter app in examples
